refactor: extract safe lambda module
